requirements.txt
fastapi
uvicorn
streamlit
tensorflow
numpy
pillow
requests
pydantic
python-multipart